move utility functions to package
