+++ src/GUIWidgets.py
@@ -26,7 +26,7 @@ from PyQt5.QtWidgets import (
 )
 from pyqtspinner import WaitingSpinner
 
-from src import utils
+from src.utils import utils
 
 
 # Create namedtuples for storing button and label configurations
